Adjust Mdl presentation code to follow new APIs from mobile-sdk-rs
